Renamed SimpleEnsemblePredictor class file
