+++ mala/network/simple_ensemble_predictor.py
